+++ andes/models/exciter.py
@@ -1,8 +1,8 @@
 from andes.core.model import ModelData, Model
-from andes.core.param import NumParam, IdxParam, ExtParam
+from andes.core.param import NumParam, IdxParam, ExtParam, dummify
 from andes.core.var import Algeb, ExtState, ExtAlgeb, State
-from andes.core.service import ConstService, ExtService
-from andes.core.block import LagAntiWindup, LeadLag, Washout, Lag
+from andes.core.service import ConstService, ExtService, VarService, PostInitService
+from andes.core.block import LagAntiWindup, LeadLag, Washout, Lag, HVGate, Piecewise, GainLimiter
 from andes.core.discrete import HardLimiter
 
 
@@ -70,7 +70,7 @@ class ExcBase(Model):
                                model='SynGen',
                                indexer=self.syn,
                                tex_name=r'X_{ad}I_{fd}',
-                               info='Armature reaction to excitation current',
+                               info='Armature excitation current',
                                )
         # from bus, get a and v
         self.a = ExtAlgeb(model='Bus',
@@ -86,23 +86,12 @@ class ExcBase(Model):
                           info='Bus voltage magnitude',
                           )
 
-        # input excitation voltages; PSS outputs summed at vi
-        self.vi = Algeb(info='Total input voltages',
-                        tex_name='V_i',
-                        unit='p.u.',
-                        )
         # output excitation voltage
         self.vout = Algeb(info='Exciter final output voltage',
                           tex_name='v_{out}',
                           v_str='vf0',
                           )
 
-        self.vref = Algeb(info='Reference voltage input',
-                          tex_name='V_{ref}',
-                          unit='p.u.',
-                          v_str='vref0',
-                          e_str='vref0 - vref'
-                          )
         # Note:
         # Subclasses need to define `self.vref0` in the appropriate place
 
@@ -112,7 +101,7 @@ class EXDC2Data(ExcBaseData):
         super().__init__()
         self.TR = NumParam(info='Sensing time constant',
                            tex_name='T_R',
-                           default=1,
+                           default=0.01,
                            unit='p.u.',
                            )
         self.TA = NumParam(info='Lag time constant in anti-windup lag',
@@ -215,6 +204,13 @@ class EXDC2Model(ExcBase):
                                   v_str='vb0 + v',
                                   )  # derived classes to-do: provide `v_str`
 
+        self.vref = Algeb(info='Reference voltage input',
+                          tex_name='V_{ref}',
+                          unit='p.u.',
+                          v_str='vref0',
+                          e_str='vref0 - vref'
+                          )
+
         self.Se = Algeb(info='Saturation output',
                         tex_name='S_e',
                         unit='p.u.',
@@ -225,14 +221,19 @@ class EXDC2Model(ExcBase):
                         tex_name='V_p',
                         unit='p.u.',
                         v_str='vf0',
-                        e_str='LA_x - KE*vp - Se*vp',
+                        e_str='LA_y - KE*vp - Se*vp',
                         t_const=self.TE,
                         )
 
         self.LS = Lag(u=self.v, T=self.TR, K=1.0, info='Sensing lag TF')
 
+        # input excitation voltages; PSS outputs summed at vi
+        self.vi = Algeb(info='Total input voltages',
+                        tex_name='V_i',
+                        unit='p.u.',
+                        )
         self.vi.v_str = 'vb0'
-        self.vi.e_str = '(vref - LS_x - W_y) - vi'
+        self.vi.e_str = '(vref - LS_y - W_y) - vi'
 
         self.LL = LeadLag(u=self.vi,
                           T1=self.TC,
@@ -310,6 +311,20 @@ class SEXSModel(ExcBase):
                                   v_str='vf0/K + v',
                                   )
 
+        self.vref = Algeb(info='Reference voltage input',
+                          tex_name='V_{ref}',
+                          unit='p.u.',
+                          v_str='vref0',
+                          e_str='vref0 - vref'
+                          )
+        # input excitation voltages; PSS outputs summed at vi
+        self.vi = Algeb(info='Total input voltages',
+                        tex_name='V_i',
+                        unit='p.u.',
+                        )
+        self.vi.e_str = '(vref - v) - vi'
+        self.vi.v_str = 'vref0 - v'
+
         self.LL = LeadLag(u=self.vi, T1=self.TA, T2=self.TB)
 
         self.LAW = LagAntiWindup(u=self.LL_y,
@@ -319,10 +334,7 @@ class SEXSModel(ExcBase):
                                  upper=self.EMAX,
                                  )
 
-        self.vout.e_str = 'LAW_x - vout'
-
-        self.vi.e_str = '(vref - v) - vi'
-        self.vi.v_str = 'vref0 - v'
+        self.vout.e_str = 'LAW_y - vout'
 
 
 class SEXS(SEXSData, SEXSModel):
@@ -337,7 +349,7 @@ class EXST1Data(ExcBaseData):
     def __init__(self):
         ExcBaseData.__init__(self)
 
-        self.TR = NumParam(default=0.02,
+        self.TR = NumParam(default=0.01,
                            info='Measurement delay',
                            tex_name='T_R',
                            )
@@ -399,6 +411,21 @@ class EXST1Model(ExcBase):
                                   v_str='v + vf0 / KA',
                                   )
 
+        self.vref = Algeb(info='Reference voltage input',
+                          tex_name='V_{ref}',
+                          unit='p.u.',
+                          v_str='vref0',
+                          e_str='vref0 - vref'
+                          )
+
+        # input excitation voltages; PSS outputs summed at vi
+        self.vi = Algeb(info='Total input voltages',
+                        tex_name='V_i',
+                        unit='p.u.',
+                        )
+        self.vi.v_str = 'vf0 / KA'
+        self.vi.e_str = '(vref - LG_y - WF_y) - vi'
+
         self.LG = Lag(u=self.v, T=self.TR, K=1,
                       info='Sensing delay',
                       )
@@ -417,7 +444,7 @@ class EXST1Model(ExcBase):
 
         self.LR = Lag(u=self.LL_y, T=self.TA, K=self.KA, info='Regulator')
 
-        self.WF = Washout(u=self.LR_x, T=self.TF, K=self.KF, info='Stablizing circuit feedback')
+        self.WF = Washout(u=self.LR_y, T=self.TF, K=self.KF, info='Stablizing circuit feedback')
 
         # the following uses `XadIfd` for `IIFD` in the PSS/E manual
         self.vfmax = Algeb(info='Upper bound of output limiter',
@@ -434,10 +461,7 @@ class EXST1Model(ExcBase):
         self.HLR = HardLimiter(u=self.WF_y, lower=self.vfmin, upper=self.vfmax,
                                info='Hard limiter on regulator output')
 
-        self.vi.v_str = 'vf0 / KA'
-        self.vi.e_str = '(vref - LG_x - WF_y) - vi'
-
-        self.vout.e_str = 'LR_x*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu - vout'
+        self.vout.e_str = 'LR_y*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu - vout'
 
 
 class EXST1(EXST1Data, EXST1Model):
@@ -448,3 +472,280 @@ class EXST1(EXST1Data, EXST1Model):
     def __init__(self, system, config):
         EXST1Data.__init__(self)
         EXST1Model.__init__(self, system, config)
+
+
+class ESST3AData(ExcBaseData):
+    def __init__(self):
+        ExcBaseData.__init__(self)
+        self.TR = NumParam(info='Sensing time constant',
+                           tex_name='T_R',
+                           default=0.01,
+                           unit='p.u.',
+                           )
+        self.VIMAX = NumParam(default=0.8,
+                              info='Max. input voltage',
+                              tex_name='V_{IMAX}',
+                              vrange=(0, 1),
+                              )
+        self.VIMIN = NumParam(default=-0.1,
+                              info='Min. input voltage',
+                              tex_name='V_{IMIN}',
+                              vrange=(-1, 0),
+                              )
+
+        self.KM = NumParam(default=500,
+                           tex_name='K_M',
+                           info='Forward gain constant',
+                           vrange=(0, 1000),
+                           )
+        self.TC = NumParam(info='Lead time constant in lead-lag',
+                           tex_name='T_C',
+                           default=3,
+                           vrange=(0, 20),
+                           )
+        self.TB = NumParam(info='Lag time constant in lead-lag',
+                           tex_name='T_B',
+                           default=15,
+                           vrange=(0, 20),
+                           )
+
+        self.KA = NumParam(info='Gain in anti-windup lag TF',
+                           tex_name='K_A',
+                           default=50,
+                           vrange=(0, 200),
+                           )
+        self.TA = NumParam(info='Lag time constant in anti-windup lag',
+                           tex_name='T_A',
+                           default=0.1,
+                           vrange=(0, 1),
+                           )
+        self.VRMAX = NumParam(info='Maximum excitation limit',
+                              tex_name='V_{RMAX}',
+                              default=8,
+                              unit='p.u.',
+                              vrange=(0.5, 10),
+                              )
+        self.VRMIN = NumParam(info='Minimum excitation limit',
+                              tex_name='V_{RMIN}',
+                              default=0,
+                              unit='p.u.',
+                              vrange=(-10, 0.5),
+                              )
+        self.KG = NumParam(info='Feedback gain of inner field regulator',
+                           tex_name='K_G',
+                           default=1,
+                           vrange=(0, 1.1),
+                           )
+        self.KP = NumParam(info='Potential circuit gain coeff.',
+                           tex_name='K_P',
+                           default=4,
+                           vrange=(1, 10),
+                           )
+        self.KI = NumParam(info='Potential circuit gain coeff.',
+                           tex_name='K_I',
+                           default=0.1,
+                           vrange=(0, 1.1),
+                           )
+        self.VBMAX = NumParam(info='VB upper limit',
+                              tex_name='V_{BMAX}',
+                              default=18,
+                              unit='p.u.',
+                              vrange=(0, 20),
+                              )
+        self.KC = NumParam(default=0.1,
+                           info='Rectifier loading factor proportional to commutating reactance',
+                           tex_name='K_C',
+                           vrange=(0, 1),
+                           )
+        self.XL = NumParam(default=0.01,
+                           info='Potential source reactance',
+                           tex_name='X_L',
+                           vrange=(0, 0.5),
+                           )
+        self.VGMAX = NumParam(info='VG upper limit',
+                              tex_name='V_{GMAX}',
+                              default=4,
+                              unit='p.u.',
+                              vrange=(0, 20),
+                              )
+        self.THETAP = NumParam(info='Rectifier firing angle',
+                               tex_name=r'\theta_P',
+                               default=0,
+                               unit='degree',
+                               vrange=(0, 90),
+                               )
+        self.TM = NumParam(default=0.1,
+                           info='Inner field regulator forward time constant',
+                           tex_name='K_C',
+                           )
+
+        self.VMMAX = NumParam(info='Maximum VM limit',
+                              tex_name='V_{MMAX}',
+                              default=1,
+                              unit='p.u.',
+                              vrange=(0.5, 1.5),
+                              )
+        self.VMMIN = NumParam(info='Minimum VM limit',
+                              tex_name='V_{RMIN}',
+                              default=0.1,
+                              unit='p.u.',
+                              vrange=(-1.5, 0.5),
+                              )
+
+
+class ESST3AModel(ExcBase):
+    def __init__(self, system, config):
+        ExcBase.__init__(self, system, config)
+
+        self.KPC = ConstService(v_str='KP * exp(1j * radians(THETAP))',
+                                tex_name='K_{PC}',
+                                info='KP polar THETAP',
+                                )
+
+        # vd, vq, Id, Iq from SynGen
+        self.vd = ExtAlgeb(src='vd',
+                           model='SynGen',
+                           indexer=self.syn,
+                           tex_name=r'V_d',
+                           info='d-axis machine voltage',
+                           )
+        self.vq = ExtAlgeb(src='vq',
+                           model='SynGen',
+                           indexer=self.syn,
+                           tex_name=r'V_q',
+                           info='q-axis machine voltage',
+                           )
+        self.Id = ExtAlgeb(src='Id',
+                           model='SynGen',
+                           indexer=self.syn,
+                           tex_name=r'I_d',
+                           info='d-axis machine current',
+                           )
+
+        self.Iq = ExtAlgeb(src='Iq',
+                           model='SynGen',
+                           indexer=self.syn,
+                           tex_name=r'I_q',
+                           info='q-axis machine current',
+                           )
+
+        # control block begin
+        self.LG = Lag(self.v, T=self.TR, K=1,
+                      info='Voltage transducer',
+                      )
+
+        self.UEL = Algeb(info='Interface var for under exc. limiter',
+                         tex_name='U_{EL}',
+                         v_str='0',
+                         e_str='0 - UEL'
+                         )
+
+        self.VE = VarService(tex_name='V_E',
+                             info='VE',
+                             v_str='Abs(KPC*(vd + 1j*vq) + 1j*(KI + KPC*XL)*(Id + 1j*Iq))',
+                             )
+
+        self.IN = Algeb(tex_name='I_N',
+                        info='Input to FEX',
+                        v_str='KC * XadIfd / VE',
+                        e_str='KC * XadIfd / VE - IN',
+                        )
+
+        self.FEX = Piecewise(u=self.IN,
+                             points=(0, 0.433, 0.75, 1),
+                             funs=('1', '1 - 0.577*IN', 'sqrt(0.75 - IN ** 2)', '1.732*(1 - IN)', 0),
+                             info='Piecewise function FEX',
+                             )
+
+        self.VBMIN = dummify(-9999)
+        self.VGMIN = dummify(-9999)
+
+        self.VB = GainLimiter(u='VE*FEX_y',
+                              K=1,
+                              upper=self.VBMAX,
+                              lower=self.VBMIN,
+                              no_lower=True,
+                              info='VB with limiter',
+                              )
+
+        self.VG = GainLimiter(u=self.vout,
+                              K=self.KG,
+                              upper=self.VGMAX,
+                              lower=self.VGMIN,
+                              no_lower=True,
+                              info='Feedback gain with HL',
+                              )
+
+        self.vrs = Algeb(tex_name='V_{RS}',
+                         info='VR subtract feedback VG',
+                         v_str='vf0 / VB_y / KM',
+                         e_str='LAW1_y - VG_y - vrs',
+                         )
+
+        self.vref = Algeb(info='Reference voltage input',
+                          tex_name='V_{ref}',
+                          unit='p.u.',
+                          v_str='(vrs + VG_y) / KA + v',
+                          e_str='vref0 - vref',
+                          )
+
+        self.vref0 = PostInitService(info='Initial reference voltage input',
+                                     tex_name='V_{ref0}',
+                                     v_str='vref',
+                                     )
+
+        # input excitation voltages; PSS outputs summed at vi
+        self.vi = Algeb(info='Total input voltages',
+                        tex_name='V_i',
+                        unit='p.u.',
+                        e_str='-LG_y + vref - vi',
+                        v_str='-v + vref',
+                        )
+
+        self.vil = Algeb(info='Input voltage after limit',
+                         tex_name='V_{il}',
+                         v_str='HLI_zi*vi + HLI_zl*VIMIN + HLI_zu*VIMAX',
+                         e_str='HLI_zi*vi + HLI_zl*VIMIN + HLI_zu*VIMAX - vil'
+                         )
+
+        self.HG = HVGate(u1=self.UEL,
+                         u2=self.vil,
+                         info='HVGate for under excitation',
+                         )
+
+        self.LL = LeadLag(u=self.HG_y, T1=self.TC, T2=self.TB,
+                          info='Regulator',
+                          )  # LL_y == VA
+
+        self.LAW1 = LagAntiWindup(u=self.LL_y,
+                                  T=self.TA,
+                                  K=self.KA,
+                                  lower=self.VRMIN,
+                                  upper=self.VRMAX,
+                                  info='Lag AW on VR',
+                                  )  # LAW1_y == VR
+
+        self.HLI = HardLimiter(u=self.vi,
+                               lower=self.VIMIN,
+                               upper=self.VIMAX,
+                               info='Input limiter',
+                               )
+
+        self.LAW2 = LagAntiWindup(u=self.vrs,
+                                  T=self.TM,
+                                  K=self.KM,
+                                  lower=self.VMMIN,
+                                  upper=self.VMMAX,
+                                  info='Lag AW on VM',
+                                  )  # LAW2_y == VM
+
+        self.vout.e_str = 'VB_y * LAW2_y - vout'
+
+
+class ESST3A(ESST3AData, ESST3AModel):
+    """
+    Static exciter type 3A model
+    """
+    def __init__(self, system, config):
+        ESST3AData.__init__(self)
+        ESST3AModel.__init__(self, system, config)