+++ andes/system.py
@@ -329,11 +329,12 @@ class System(object):
         # set internal variable addresses
         for mdl in models.values():
             if mdl.flags['address'] is True:
-                logger.debug(f'{mdl.class_name:10s}: addresses exist.')
+                logger.debug(f'{mdl.class_name} address exists')
                 continue
             if mdl.n == 0:
                 continue
 
+            logger.debug(f'Setting address for {mdl.class_name}')
             n = mdl.n
             m0 = self.dae.m
             n0 = self.dae.n
@@ -468,6 +469,8 @@ class System(object):
 
             self.vars_to_dae(mdl)
             self.vars_to_models()
+
+        self.s_update_post(models)
 
         # store the inverse of time constants
         self._store_tf(models)
@@ -608,6 +611,23 @@ class System(object):
         """
         self.call_models('l_check_eq', models)
         self.call_models('l_set_eq', models)
+
+    def s_update_var(self, models: Optional[Union[str, List, OrderedDict]] = None):
+        """
+        Update variable services by calling ``s_update_var`` of models.
+
+        This function is must be called before any equation evaluation after
+        limiter update function `l_update_var`.
+        """
+        self.call_models('s_update_var', models)
+
+    def s_update_post(self, models: Optional[Union[str, List, OrderedDict]] = None):
+        """
+        Update variable services by calling ``s_update_post`` of models.
+
+        This function is called at the end of `System.init()`.
+        """
+        self.call_models('s_update_post', models)
 
     def fg_to_dae(self):
         """